requirements.txt
Django>=4.2
djangorestframework>=3.14.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
PyJWT>=2.7.0
passlib[bcrypt]>=1.7.4
asgiref>=3.7.2
bcrypt>=4.0.1
tensorflow>=2.11
torch>=2.0.0        # optional if you won't use torch; safe to have
scikit-learn>=1.2
xgboost>=1.7
pillow>=9.0
numpy>=1.23
pandas>=1.5
joblib>=1.2
matplotlib>=3.5
tqdm>=4.65
streamlit
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
joblib
pillow
numpy
tensorflow   # or torch depending on your model; include only what you need
scikit-learn
xgboost